internal/v1: split out remaining handler tests
This way server_test.go is just the helper functions.
